28.08.18: Виправив внесену помилку при розрахунку 3I0 по первинній  обмотці

+++ working files/src/protections.c
@@ -1294,7 +1294,7 @@ inline void calc_measurement(unsigned int number_group_stp)
     ortogonal_local_3I0[1] = ((MNOGNYK_3I0_DIJUCHE_D_mA*ortogonal_calc[2*FULL_ORT_3I0 + 1]) << (VAGA_DILENNJA_I_DIJUCHE - (VAGA_DILENNJA_3I0_DIJUCHE_D_mA + 4)))/MNOGNYK_I_DIJUCHE;
 #endif
   
-    unsigned int T0 = current_settings_prt.T0, TCurrent = current_settings_prt.TCurrent;
+    int T0 = (int)current_settings_prt.T0, TCurrent = (int)current_settings_prt.TCurrent;
     _x = ortogonal_calc[2*FULL_ORT_Ib + 0] = T0*ortogonal_local_3I0[0]/TCurrent - (ortogonal_calc[2*FULL_ORT_Ia + 0] + ortogonal_calc[2*FULL_ORT_Ic + 0]);
     _y = ortogonal_calc[2*FULL_ORT_Ib + 1] = T0*ortogonal_local_3I0[1]/TCurrent - (ortogonal_calc[2*FULL_ORT_Ia + 1] + ortogonal_calc[2*FULL_ORT_Ic + 1]);
     if (copy_to_low_tasks == true)
